Ladybug.0.0.53
Multiple minor modifications. This version is a pre-release of Ladybug
that works with Honeybee. I removed the zip file from the github not to
duplicate the stable copy. The zip file will be only available on the
box folder for download.
The structure of the repo is cleaner now just .\src and .\userObjects
What is really special about this commit is the fact that all the source
code and userObjects are generated automatically by a single component.
Next step is to finish the component that sync the files with github
repo.