Made a new tab for WIP components
Now that multiple developers are developing for Ladybug I added this tab
where we can keep put the new components to be tested by the users
before add them to Ladybug.
I moved the new solar fan and solar envelope to this tab and will ask
more users to test it out.