Modified sunpath to work with list of HOY
It has been a while that I wanted to modify the sunpath code so instead
of iterating through days. months and hours which has multiple
limitations the component calculates the hours of the year based on the
inputs for months, days, hours or analysis period and then iterates
through them.