BRAND NEW SHADE BENEFIT EVALUATOR!!!!
Introducing the Ladybug Shad Benefit Evaluator!

This is a component for visualizing the desirability of shading over a
window by using solar vectors, the outdoor dry bulb temperature, and an
assumed building balance temperature.

Solar vectors for hours when the temperature is above the balance point
contribute positively to shade desirability while solar vectors for
hours when the temperature is below the balance point contribute
negatively.

The component outputs a colored mesh of the shade illistrating the net
effect of shading each mesh face.  A higher saturation of blue indicates
that shading the cell is very desirable.  A higher saturation of red
indicates that shading the cell is harmful (blocking more winter sun
than summer sun). Desaturated cells indicate that shading the cell will
have relatively little effect on building performance.

The method used by this component is based off of the Shaderade method
developed by Christoph Reinhart, Jon Sargent, Jeffrey Niemasz.  A
special thanks goes to them and their research.  A paper detailing the
shaderade method is available at:

http://www.gsd.harvard.edu/research/gsdsquare/Publications/Shaderade_BS2011.pdf