SHADING BENEFIT READY FOR USER TESTING!
Final fixes to the component before user testing:

Added in better descriptions of the inputs, outputs and process the
component runs so that users can understand what's going on.
Added a few more checks to ensure that the component does not fail.
Fixed an issue with the colors and default legend generation.
Developed a better way to generate test points over oddly-shaped window
or test region geometry.
Added in an input for context that will discount solar vectors that get
blocked by such geometry from the calculation.