Massive Component Descriptions Update Part 1
Updated all of the text descriptions and input/output descriptions for
roughly half of the Ladybug components.  Also, all of the component
warnings and the "preview in Rhino" options have been standardized to
make Ladybug amenable to new users of the suite.  Lastly, a few bugs
were fixed in the sky mask component and some changes made to the
outputs to make it more user-friendly.