Massive Component Descriptions Update Part 2
Last and final update for all of the descriptions of components inputs
and outputs.  The GH component warnings and preview options of the
components have also been standardized and synchronized.  Ladybug is now
finally in a state where a new user might be able to figure out how the
components work by reading the component descriptions instead of
searching online.