Import Stat Component Brings in Typical Periods
The import stat component has been modified such that it now reads the
typical and extreme fields of the stat file and brings them into GH as
analysis periods.  The code was made to be adaptable to all of the
different levels of detail of the periods for different stat files.