New Open EPW and STAT Together
A component that takes as input a URL of a DOE zip file and
automatically downloads and unzips the file and brings both the EPW and
STAT files into Grasshopper.

Also, I fixed a few typos in the Import Stat component as well as
writing some better descriptions for the Bounce from Surface Component.
It took me a long time to understand how the component works and I
wouldn't want anyone else to have to go through it.