Fixed Error in Fan/Envelope/SHDDesigner and WIP
An error has been fixed in the solar fan that was causing the fan to
fail when someone put in vectors that were almost parallel to the test
surface.  The fan now works in most cases and all shading components
will throw an error if a vector is so close to being parallel that it is
within the angle tolerance.

Also included in this commit is a Shading Evaluator for the WIP tab.