Download EPW/STAT With Extra Checks
The component will now check to see if the user has already downloaded a
certain file to C:\ladybug before trying to download the file.  The
download function now also has a check, which will throw a warning
message if the user is not connected to the internet (instead of before
when the component would just fail).