3D Chart Makeover
The 3D chart has been revamped with the following awesome new features:
1) The mesh is now generated with one solid-color face per each hour.
This makes it much clearer to see when certain phenomena are happening.
2) Output curves and time labels have been added using the analysis
periods in the header.  This makes it much easier to see at which months
of the year and hours of the day certain phenomena are happening.  These
curves and text surfaces will not be generated if the analysis period
cannot be read from the header or if there are not 8760 values in the
input list.
3) The component now outputs text strings for the title and time labels
along with base points for this text such that users can hook them up to
the TextTag3D if they so desire.
4) Since having a toggle to delete vertices based on a conditional
statement was confusing some people, now, whenever someone hooks up a
conditional statement, the mesh faces that do not meet the conditions
will automatically be removed.