Fixed a Bug in the GetHOYs Function
There was a bug in the GetHOYs function based on the analysis period.
The function was forgetting to include midnight.  I originally copied
the function from the sun path component and I imagine that the bug was
not important because there is usually no midnight sun.  We should
probably fix this in case we get a weather file for the arctic.