Fixed Bugs In Depth Method
A number of bugs were fixed in the depth method including a series of
checks that make sure that the correct number of shades is generated
starting from either the top in the case of horizontal shades or from
the northernmost side in the case of vertical shades.  Also, the shade
angle now works when there is no input for HorOrVertical.