SOLAR TEMPERATURE ADJUSTOR IS HERE!!!!
I am happy so say that we can now adjust mean radiant temperature for
shortwave solar radiation in a manner that is reliable.

The component combines radiance functions over a comfort mannequin to
estimate the change in MRT for a human sitting outside in the sun.

This is hopefully the first of many components that will adjust MRT.
The next ones will take into account proximity to surfaces simulated
with E+.

Note that this component currently takes a long time to adjust MRT for a
whole year and it is currently only working for Rhino docs in Meters.  I
will correct this for other unit systems tomorrow.