PSYCHROMETRIC CHART COMPLETE!!!
The psychrometric chart is, at last, complete in terms of all of the
standard features of ladybug components.  You can now hook up
annualhourly data, put in conditional statements, and get out colors for
points that correspond to the hourly data.

Finally, you can adjust the assumptions of the strategy polygons with a
new strategyPar component and some of the characteristics of comfort
that are scientifically contested have been moved into a separate
comfortPar component.