SKY COLOR VISUALIZER IS HERE!!!!
This is an initial commit of a component to visualize the color of the
sky dome using the functions of @twastvedt..  The code can be edited
easily by editing the skyColor class in Ladybug Ladybug.  As it stands,
now the component is mostly meant to compliment the sunPath to show the
qualities of the sun at a given hour.  The sky color mesh can also be
unrolled into a rectangle with a boolean option.