Solar Temp Calculator Runs at 12 Times Speed
I have gotten the solar temperature adjustor to run at 12 times the
speed by 1) first checking the sun-up hours of the year and not looking
as skies where there is not sun and 2) running the calculation for each
sky in parallel.

Also, I discovered a bug in the way @mostaphaRoudsari made the shade
benefit evaluator run in parallel.  It has to do with the fact that,
when you have a "for" loop run in parallel, the items will be appended
to the list in the order in which they finish calculating.  Not the
order of the original list that they are running through.  I will try to
fix this bug tomorrow.