Re-Vamped Solar Temperature Adjustor
I put in a bunch of methods to run very fast approximations of
solar-adjusted MRT based on the paper by Tyler Hoyt.  I still have two
variables to touch-up a bit in the fastest possible method but there is
now a simplified mannequin mesh that will run things very fast.  Also,
the context geometry is now working such that I can begin to incorporate
this into the indoor MRT maps.