FINISHED NEW IMPROVED SOLAR CALCULATOR!
The new method is super-fast and gives number that are pretty close to
the more accurate simulations with the MRT mannequin.  I have to hand it
to the CBE team for making such an awesome fast way to calculate this.
Before, the annual calculation of MRT delta used to take about 7 minutes
at least (in parallel).  Now, the whole year comes out in 25 seconds
with no need for a parallel calculation!