Major Text Format Update
After getting the opinion of a few users, I realized that no one uses
the outlines of the text surfaces in any vector-based workflows like
exporting to Illustrator.  In those cases, people just completely
re-make the text in Illustrator.  So I realized that our text options
should be optimized for the display in the Rhino window for quick
workflows.  The default display of grasshopper Breps with outline curves
has caused the text to be really unclear in a lot of cases where you
zoom out of the graphic.  Accordingly, I thought the the best move was
to have the text2Srf function output black meshes for text and this
ensure that the text reads the clearest in the Rhino display.
In case people liked a bit of the old outline effect, I added the option
to bold the text in the LegendParameters component.
Enjoy!