Ladybug Wind Components Ready for Release
Finally, thanks to the code that @stgeorges developed for his original
2D wind profile and a recent combined 2D/3D component, I believe that I
can say that the wind profile component is ready to be really helpful to
nat vent experts.  The commit here includes almost all of the recent
code development by @stgeorges and a number of edits that I have made to
apply the awesome code of the 2D wind profile to the 3D profile.  I will
check tomorrow with @stgeorges if all of the edits are ok by him and it
looks like the users are in for a real treat with the next release!