Radiation Studies  now in kWh and Other Good Stuff
A lot of stuff improved. I have ensured that all meshes of visualization
components are facing the right the direction in the event that the GH
display goes weird and only shows meshes in one direction.  I have also
improved a bunch of things with the radiation components, notably the
radiation rose (with a new function in rhinocommon instead of rs that
does not fail to generate arrows with small analysis periods).  I also
smoothed out the colors of the Cala lily and added in the option to make
a dome.  Lastly, I put esc key functionality into the MRT component.  I
noticed that the method I was using only works for non-parallel
calculations.  I will have to figure out how to cancel for parallel
calculations.