Introducing Adaptive Comfort Parameters!
Ladybug now includes a range of ways to adjust the adaptive model.
Notably, Ladybug now has support for both American and European adaptive
standards and even allows users to combine methods of the two standards
(with appropriate comments given, of course).

This commit also introduces the ability to make use of research that has
applied the adaptive methodology to conditioned buildings (with a
levelOfConditioning input).

Next on the order of business, I will finish of the Adaptive Comfort
Chart and get it out of WIP so that we all have a means of visualizing
these different standards.