ADAPTIVE COMFORT CHART IS HERE!
Finally!
I can say that all of the features of the adaptive comfort chart are in
place and it is ready to emerge from its shell in WIP.  Making this
component also really helped me think through what to do in the
scenarios that are not officially covered by the ASHRAE or European
standards so that the adaptive calculations on all of the other adaptive
components should make more sense now.

Aside from one final homage to Europe with the Molier HX option in the
psych chart, this commit concludes my work with thermal comfort models
in LB+HB.

It's been a good long run and I am glad that I have seen it through to
the end!