Gave Ground Temp Nicer Icon
I did not originally intend for all of the ground temp stuff to be in
one component. I had a nicer icon than the one I had given to Anton
ready for a ground visualization component.  I thought that I should put
in this nicer icon on the component.  Also, since the component is
really about visualizing the ground temperature now, I have moved it to
the Visualize Weather Data tab.