Cleaned Up Code
I know @antonszilasi , that you were just learning when you wrote this
but the way that the functions were structured was making it completely
unreadable.  There were also two instances of duplicate functions (each
of which were very long).  I hope that you don't mind that I
re-organized the code as a series of functions followed by a main chunk
of code at the bottom and got rid of the duplicate functions.