Added Components to Convert by 1/1000 factor
@arkdanielnielsen ,
I know that you had plans to implement something like this and I am
sorry for taking your thunder here.  The need for this component just
became painfully obvious as I started working for an architecture office
and so I have made the component for now.  Feel free to edit or make any
other converters (I am realizing that there are still a lot that we
need).