Solar Adjusted Temp Now Accounts for Sky Temperature
@stgeorges ,
I realized that the reason why our components were giving such different
answers for MRT as noted in this discussion:

http://www.grasshopper3d.com/group/ladybug/forum/topic/show?id=2985220%3ATopic%3A1520583&xg_source=msg
It was because you were accounting for all long wave radiative transfer
and my component was only built to take an existing long wave MRT and
adjust it for short wave radiation.  As such, the original solar
adjusted component was only meant to be used in cases where we have an
existing long wave MRT, which usually only happens when we run an indoor
energy simulation.  The fact that the SolarCal method only accounts for
short wave radiation is also the main reason why the authors of SolarCal
stated that they were mking a model primarily for indoor conditions.
With this commit, I have added in the ability to run a long wave
radiation calculation inside the solar adjusted component if the user
connects up outdoor dry bulb temperature.  I am using the MENEX model to
compute long wave sky temperature.  When used in this capacity, the
results of the solar adjusted temperature component are aligning with
you thermal indices component well.  The only times when they appear to
deviate is when solar radiation becomes particularly intense, which I
think is a result of the fact that the MENEX model uses some relatively
simple assumptions about human geometry.