Added Horizontal Infrared Radiation to Import EPW
I realized that both @stgeorges and I could really use this data in a
number of our components and I will post a full list of reasons here:

http://www.grasshopper3d.com/group/ladybug/forum/topics/difference-between-mrt-calculation-methods?commentId=2985220%3AComment%3A1537960&groupId=2985220%3AGroup%3A658987

Also, in an attempt to not mess up all of the other components that
reference the epwDataReader function, I have taken out
liquidPrecipitationDepth.  Even without this particular reason, however,
I was thinking that we should remove this precipitation output as there
are almost no EPW files that have this information and those that do
have it usually have incorrect data that does not align with annual
average precipitation.  In the future, I imagine us pulling this
information from NASA databases, which will be much more accurate.  If
anyone objects to this removal, please comment on this commit with
reasons and I can find a way to work it back in.