updates to "Sunpath shading", "Terrain Generator 2", "Terrain analysis" components (#300)
* Delete Ladybug_Terrain Generator 2.ghuser

* Add files via upload

* Fixing something with creation of type_ = 2 and type_ = 3

Setting the "loftedTerrain_Outline_and_OutlineProjected_Brep" to always be the first face of the "terrainUnoriginUnscaledUnrotated". Fixing some minor bugs (adding: ```downloadUrl_ = None```).

* adding new "Terrain analysis" and "Flow paths" component's .ghuser files

* adding new "Terrain analysis" and "Flow paths" component's source files

* Delete Ladybug_Sunpath Shading.ghuser

* Delete Ladybug_Terrain Generator 2.ghuser

* Delete Ladybug_Terrain Analysis.ghuser

* Add files via upload

* fixing "Sunpath shading" component's issue with vertical PVsurface tolerance

This commit fixes the tolerance issue described on forum:
http://www.grasshopper3d.com/xn/detail/2985220:Comment:1610876

* rename "Terrain Generator 2" component's "originPt" output to "origin"

This change is a result of a discussion with Antonello on forum:
http://www.grasshopper3d.com/group/ladybug/forum/topics/lb-terrain-generator?commentId=2985220%3AComment%3A1610888&xg_source=activity&groupId=2985220%3AGroup%3A658987

* updates to "Terrain Analysis" component

renaming the "_originPt" input to "_origin" as a result of a discussion with Antonello on forum:
This change is a result of a discussion with Antonello on forum:
http://www.grasshopper3d.com/group/ladybug/forum/topics/lb-terrain-generator?commentId=2985220%3AComment%3A1610888&xg_source=activity&groupId=2985220%3AGroup%3A658987

Fixing the bug with visibility analysis type coloring, and a few more small bugs.