Implmented Stereographic Projection on Cala Dome
... also, I can't belive that it took me this long to realize this,
@mostaphaRoudsari , but I finally figured out how the Cala Lily is
supposed to be read and I realized that I have been reading it backwards
for 3 years now.  The fact that the tallest part of the sky appears on
the edges of the circle was really counter-intuituve to me.  Given that
I have edited the code of this component on several occaisions and still
completely misinterpreted it, I  have to imagine that many other users
have also misinterpreted it.  A lot of things became clear once I viewed
the stereographic projection of the hemisphere in relation to the cala
Lilly.  As a result, I hope you don't mind that I have renamed the
component to just be the "Cala Dome" and taken out the Lily option.  I
feel that this move is justified given that the new stereograhpic
projection of the cala dome is essentailly a more intutive way of
visualizing the original lily and so I feel that I am just replacing one
feature with another (not taking anything out).