Rewrote and Improved Shading Mask
Taking into account a lot of code and suggestions from @Vyrwn .  It's
now much faster with more accurate sky view/exposure calculations now
that we have freed ourselved of the need to use Tregenza sky patches
(using just the mesh divisions).  The resulting mask also looks smoother
and more beautiful.  Finally, I have incorporated the suggestion of a
coworker who wanted to use the mask to visualize the sun blocked by an
overhang or vertical fin with a given projection angle.  This should
increase the shading mask's usability with the new stereographic
sunpath.