Added a new component to update files
This component addresses a couple of issues that we have with the
current updated components and is designed to be used by users.

1. It separates updating the installation from updating the file.
2. It updates all the components for honeybee, ladybug, butterfly or
dragonfly as far as there is a newer version of the component is
available as a userobject.
3. It marks the components that can't be updated automatically because
of the changes in input / output names.
4. It inserts a new updated component next to the component in case of 3
so all user needs to do is to insert the connect the new component.

I think we had another issue for this opened by @ayezioro but i can't
find it so just check this discussion:
http://www.grasshopper3d.com/group/ladybug/forum/topics/honeybee-updatethisfile-error
Resolved #303