Setting Default Z Scale of 3D Chart to 0
After doing this manually every single time that I use the chart, I have always left the z-scale as non-zero because I think it's a nice way to teach people about how to edit the component.  However, the thing tha has pushed me over the edge is that Rhino 6 treats a lot of these 3D meshes as invalid.  So I'm changing it now.