Avoiding Div/0 in MeshSelector (#464)
* Avoiding Div/0

Avoiding Div/0 when no point meets levelOfPreform.  This is important if you have more than one analysisResult because otherwise the script stops at div/0 and other results aren't evaluated

* Avoiding Div/0

Avoding Div/0 if none of the TestPoints meets the PerformanceLevel.

* Revert "Avoiding Div/0"

This reverts commit fe6de3a3b30a2f7c21acf46c9184c57e4fe42e7d.

* Revert "Avoiding Div/0"

This reverts commit 02c9c78761ec75b3975746c4030d049a37fd17e2.

* Avoiding Div/0

Avoiding Div/0 when none of the TestPoints meet the levelOfPerformance. This is necessary because otherwise the script stops and doesn't evaluate further resultets

* Avoid Div/0

Avoiding Div/0 when none of the TestPoints meet the levelOfPerformance. This is necessary because otherwise the script stops and doesn't evaluate further resultets

* Avoid Div/0

This time with changed code & the correct date :-)